Slide: Rectangle as root component, for bg customization
I am writing a presentation using the default openbossa theme.
It is an experiment aiming to find what would be a good API and
set of components to create later a presentation tool itself.
I changed the root component to Rectangle so I can change the background
color. The default font color was changed to white just because the
default bg is black.

Signed-off-by: Anselmo L. S. Melo <anselmo.melo@openbossa.org>