Presentation: Rectangle
Also modified the root component of Presentation, due to the same
reason written in the previous commit.
There is another point: in function switchSlides, the Presentation bg
color is set to the same color of the Slide. I introduced this line to
check if we can continue with the current approach of having a limited
area for the slide, smaller than the presentation.

Signed-off-by: Anselmo L. S. Melo <anselmo.melo@openbossa.org>